mdc/backend/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
Django==3.1
django-celery-beat==2.0.0
django-cors-headers==3.5.0
django-timezone-field==4.0
djangorestframework==3.11.1
django-sendgrid-v5==0.9.0
install==1.3.3
kombu==4.6.11
nodejs==0.1.1
optional-django==0.1.0
Pillow==7.2.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
psycopg2==2.8.5
pytz==2020.1
redis==3.5.3
sendgrid==6.4.7
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
